requirements.txt
fastapi
sqlalchemy
alembic
psycopg2
python-dotenv
uvicorn
PyJWT
bcrypt
passlib
langchain
langchain-community
langchain-groq
yt_dlp
openai-whisper
nltk
groq
numpy
youtube-transcript-api
deepmultilingualpunctuation
sentence-transformer
